# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi API",
        "description": "API documentation for the Presensi application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "/docs"
        }
    ],
    "paths": {
...
        "/yf/callplan/data-callplan": {
            "get": {
                "tags": [
                    "Kunjungan"
                ],
                "summary": "Mendapatkan data jadwal kunjungan",
                "description": "Mendapatkan data karyawan yang sedang login berdasarkan token autentikasi",
                "operationId": "efa1d36023a8ae5b65a9321a65e59f1e",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data kunjungan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "employee_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "employee_email": {
                                                    "type": "string",
                                                    "example": "johndoe@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Jadwal Kunjungan Tidak tersedia diperiode ini ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Employee data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}